requirements.txt
datapizza-ai>=0.0.2
datapizza-ai-clients-google
google-generativeai>=0.3.0
qdrant-client>=1.7.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
mcp>=1.0.0
